Restore unit test timeout limit of 500ms
By using golang build tags we can separate the generative, acceptance
and unit tests and allow individual settings, like timeout.

This restores the 500ms unit test timeout limit, moves the 30s limit to
the generative test.
