+++ .golangci.yaml
@@ -13,6 +13,12 @@
 # limitations under the License.
 
 ---
+run:
+  build-tags:
+    - acceptance
+    - generative
+    - integration
+    - unit
 linters:
   enable:
     - goimports
