+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "go.inferGopath": false
+    "go.inferGopath": false,
+    "go.buildFlags": ["-tags=acceptance,generative,integration,unit"]
 }
