+++ internal/acceptance/testenv/testenv_test.go
@@ -14,6 +14,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build unit
+
 package testenv
 
 import (
