+++ internal/ec_errors/ec_error_test.go
@@ -14,6 +14,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build unit
+
 package ec_errors
 
 import (
