+++ internal/evaluation_target/application_snapshot_image/application_snapshot_image_test.go
@@ -14,6 +14,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build unit
+
 package application_snapshot_image
 
 import (
